bytebrain-server/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "ByteBrain: Chat with Your Documents!"
authors = ["Milad Khajavi"]

[tool.poetry.dependencies]
python = "3.10.12"
fastapi = "^0.95.1"
langchain = "0.0.344"
openai = "1.3.7"
tiktoken = "0.3.3"
chromadb = "0.4.10"
unstructured = "0.5.13"
youtube-transcript-api = "^0.6.0"
pytube = "^15.0.0"
google-api-python-client = "^2.88.0"
discord = "^2.2.3"
gitpython = "^3.1.31"
beautifulsoup4 = "^4.12.2"
structlog = "^23.1.0"
prometheus-client = "^0.17.1"
chat-exporter = "^2.6.1"
pysqlite3-binary = "^0.5.2"
faiss-cpu = "^1.7.4"
html2text = "^2020.1.16"
weaviate-client = "^3.24.2"
fastapi-utils = "^0.2.1"
libmagic = "^1.0"
python-magic = "^0.4.27"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
config = "config:configure"
configure_prometheus = "prometheus:configure"
webserver = "core.bots.web.webservice:main"
discord = "core.bots.discord.discord_bot:main"
main = "main:start"

dev_three_chain_step = "dev.three_chain_step:run"
dev_two_chain_step = "dev.two_chain_step:run"
dev_final_solution = "dev.final_solution:run"
download_discord_channel = "dev.discord_history:main"
crawl = "dev.crawl:main"
chromadbtest = "dev.chromadb:run"
discord_history = "dev.discord_history:main"

index_zio_project_docs = "index.index:index_zio_project_docs"
index_zionomicon_book = "index.index:index_zionomicon_book"
index_zio_project_source_code = "index.index:index_zio_project_source_code"
index_all = "index.index:index_all"
index_youtube_video = "index.index:index_youtube_video"
index_zio_ecosystem_source_code = "index.index:index_zio_ecosystem_source_code"
get_zio_ecosystem_repo_info = "index.index:get_zio_ecosystem_repo_info"

[tool.poetry.group.dev.dependencies]
deptry = "^0.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
